The Figure Update 2.0
The new version of the Figure Mode, which begins to use the Watson
Editor API to insert code into the page. This version also allows for
the insertion of a sandbox into a page. Code however has not been added
to the programs yet.

Known bugs:
Code tables don't expand after insertion.